Start tornado server that has cluster
