+++ daskdev-sat/Dockerfile
@@ -1,5 +1,5 @@
 # Daskdev-sat image for dask cluster pods
-FROM daskdev/dask:2.8.1
+FROM daskdev/dask:2.11.0
 
 RUN apt-get -qq update && \
     apt-get -qq install --yes --no-install-recommends locales > /dev/null && \
@@ -9,13 +9,12 @@ RUN apt-get -qq update && \
     rm -rf /var/lib/apt/lists/*
 ENV SHELL /bin/bash
 
-RUN conda install -y dask=2.8.1  distributed
+RUN conda install -y -c conda-forge distributed tornado kubernetes dask-kubernetes python-kubernetes
 RUN conda install -y -c https://conda.saturncloud.io/pkgs sutils=0.5.1
-RUN conda install -y -c conda-forge kubernetes=1.13.5 dask-kubernetes=0.10.0 python-kubernetes=9.0
 
 #TODO - remove this pdc crap
 #TODO - run as non-root user
 RUN mkdir -p /root/saturn/pdc/dask
-COPY dask_kubernetes_start.py /root/saturn/pdc/dask
+COPY app.py /root/saturn/pdc/dask
 
 ENV PYTHONPATH=/root/saturn
